Completed Task 1: Guess The Number.
Task-1 -Guess The Number By Giving Clue - Batch-44
Implemented 'Guess the Number' game in C++. Basic functionality includes generating a random number for the user to guess, providing feedback on whether the guess is too high or too low, and displaying the number of attempts it took to guess correctly. Code is ready for review.
c ++  programming.